move the dto in model package
