+++ backend/src/main/java/capstone/myfinancemanager/manager/service/UserServiceImp.java
@@ -1,10 +1,10 @@
 package capstone.myfinancemanager.manager.service;
 
 
-import capstone.myfinancemanager.manager.dto.UserDto;
 import capstone.myfinancemanager.manager.exceptions.PasswordNotMatchException;
 import capstone.myfinancemanager.manager.exceptions.UserExistsException;
 import capstone.myfinancemanager.manager.model.User;
+import capstone.myfinancemanager.manager.model.dto.UserDto;
 import capstone.myfinancemanager.manager.respository.UserRepo;
 import org.springframework.stereotype.Service;
