+++ backend/src/test/java/capstone/myfinancemanager/manager/service/UserServiceImpTest.java
@@ -1,9 +1,9 @@
 package capstone.myfinancemanager.manager.service;
 
-import capstone.myfinancemanager.manager.dto.UserDto;
 import capstone.myfinancemanager.manager.exceptions.PasswordNotMatchException;
 import capstone.myfinancemanager.manager.exceptions.UserExistsException;
 import capstone.myfinancemanager.manager.model.User;
+import capstone.myfinancemanager.manager.model.dto.UserDto;
 import capstone.myfinancemanager.manager.respository.UserRepo;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
